+++ react/components/Tooltip/Tooltip.types.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+interface ITooltipProps {
+  label: string;
+  children: React.ReactNode;
+}
+
+export default ITooltipProps;